@@ -6,7 +6,7 @@
 
 	<groupId>me.bvn13.kafka.health</groupId>
 	<artifactId>kafka-health-check</artifactId>
-	<version>1.5.6-SNAPSHOT</version>
+	<version>2.0.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
 	<name>Kafka Health Check</name>
@@ -16,14 +16,14 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-dependencies</artifactId>
-		<version>2.7.1</version>
+		<version>3.3.1</version>
 		<relativePath />
 	</parent>
 
 	<properties>
 		<!-- Java -->
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>16</maven.compiler.source>
+		<maven.compiler.target>16</maven.compiler.target>
 
 		<!-- Encoding -->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -31,7 +31,7 @@
 
 		<!-- Versions -->
 		<nexus.url>https://s01.oss.sonatype.org</nexus.url>
-		<guava.version>30.1.1-jre</guava.version>
+		<guava.version>32.0.0-jre</guava.version>
 		<maven-deploy-plugin.version>2.8.2</maven-deploy-plugin.version>
 		<nexus-staging-maven-plugin.version>1.6.13</nexus-staging-maven-plugin.version>
 		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
@@ -39,6 +39,13 @@
 	</properties>
 
 	<dependencies>
+
+		<dependency>
+			<groupId>jakarta.annotation</groupId>
+			<artifactId>jakarta.annotation-api</artifactId>
+			<version>3.0.0</version>
+		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-actuator</artifactId>
